refactor : remove package content
